src/main/java/com/compony/finalproject/service/impl/UserServiceImpl.java
package com.compony.finalproject.service.impl;

import com.compony.finalproject.model.User;
import com.compony.finalproject.repository.UserRepository;
import com.compony.finalproject.service.CRUDService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.NoSuchElementException;

@Service
public class UserServiceImpl implements CRUDService<User> {

    private final UserRepository userRepository;

    @Autowired
    public UserServiceImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public User getById(Long id) {
        return userRepository.findById(id).orElseThrow(NoSuchElementException::new);
    }

    @Override
    public List<User> getAll() {
        return userRepository.findAll();
    }

    @Override
    public void create(User item) {
        userRepository.save(item);
    }
}
